requirements.txt
feedparser>=6.0.10
pandas>=2.2.0
requests>=2.31.0
python-dateutil>=2.8.2
tabulate>=0.9.0
google-genai>=1.0.0
google-adk>=0.3.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
numpy>=2.0.0
